(chore): add extract text plugin for css change build directory to dist and delete src directory
