+++ __tests__/server/controllers/search.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
-import app from '../../../src/server';
+import app from '../../../dist/server';
 import models from '../../../server/models';
 
 const { Document, Role, User } = models;
