+++ client/src/components/App.js
@@ -3,6 +3,8 @@ import React, { Fragment } from 'react';
 import Main from './Main';
 import Header from '../containers/Header';
 
+import '../styles/App.scss';
+
 const App = () => (
   <Fragment>
     <Header />
