+++ client/src/styles/App.scss
@@ -1,0 +1,3 @@
+body {
+  margin: 0;
+}
