+++ package.json
@@ -38,6 +38,7 @@
     "@babel/plugin-proposal-class-properties": "^7.0.0-beta.40",
     "@babel/preset-env": "^7.0.0-beta.40",
     "@babel/preset-react": "^7.0.0-beta.40",
+    "autoprefixer": "^8.0.0",
     "axios": "^0.17.1",
     "babel-loader": "^8.0.0-beta.0",
     "bcrypt": "^1.0.3",
@@ -47,6 +48,7 @@
     "dotenv": "^4.0.0",
     "express": "^4.15.3",
     "express-validator": "^3.2.0",
+    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.6",
     "gulp": "^3.9.1",
     "gulp-babel": "^8.0.0-beta.1",
@@ -66,6 +68,7 @@
     "nodemailer-express-handlebars": "^2.0.0",
     "pg": "^6.4.0",
     "pg-hstore": "^2.3.2",
+    "postcss-loader": "^2.1.0",
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
